chore: Use project references.
Project references ensure that builds in a dependency are always up-to-date.

This provides a considerable speed up to re-builds.

All packages are now "composite" -- which is basically how we were
already building them, but it's required to set that option to reference
a project -- this allows typescript to enforce output files are
up-to-date via timestamp verification.

This patch also enables the declarationMap option which causes the
declaration file to have a source map that references back to the
original file. This means VS Code navigation within the workspace is
completely seamless.

This patch also adds toolchain pins for use with notion.

It's no longer necessary to clean the dist directory with every build
because typescript keeps it up-to-date.