feat: Ability to specify identifier ranges.
This ability lets the output of opticss on different stylesheets
co-exist in the same build by allowing configuration to specify an
identifier range for opticss to use. The build integration must
guarantee that the identifier ranges are non-overlapping.

When this is enabled, if more identifiers are needed than are available
an error is raised.